Split out concat tests into test_combine.py
