+++ xray/core/combine.py
@@ -90,7 +90,7 @@ def concat(objs, dim=None, data_vars='all', coords='different',
                       FutureWarning, stacklevel=2)
         dim = 'concat_dims'
 
-    if indexers is not None:
+    if indexers is not None:  # pragma: nocover
         warnings.warn('indexers has been renamed to positions; the alias '
                       'will be removed in a future version of xray',
                       FutureWarning, stacklevel=2)
